+++ modernization/src/Essco.Web/Pages/Account/AccessDenied.cshtml.cs
@@ -1,0 +1,7 @@
+using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Mvc.RazorPages;
+
+namespace Essco.Web.Pages.Account;
+
+[Authorize]
+public sealed class AccessDeniedModel : PageModel;